Move locale_convert into libdcp.

+++ src/lib/colour_conversion.cc
@@ -152,7 +152,7 @@ ColourConversion::as_xml (xmlpp::Node* node) const
 		in_node->add_child("B")->add_child_text (raw_convert<string> (tf->B ()));
 	}
 
-	node->add_child("YUVToRGB")->add_child_text (raw_convert<string> (_yuv_to_rgb));
+	node->add_child("YUVToRGB")->add_child_text (raw_convert<string> (static_cast<int> (_yuv_to_rgb)));
 	node->add_child("RedX")->add_child_text (raw_convert<string> (_red.x));
 	node->add_child("RedY")->add_child_text (raw_convert<string> (_red.y));
 	node->add_child("GreenX")->add_child_text (raw_convert<string> (_green.x));
